Calculate wind chill again.
In the source of http://www8.tfe.umu.se/weather-new/ is a comment
mentioning not using the value from the web service, but to calculate it
using the formula in this commit. It doesn't seem to mention where this
formula comes from or why the value from the web service is bad.